Breach Data

Simulated 7-day quarantine breach rate data across Australian states and territories, based on the modelling by Zachreson et al. (2022).

Usage

breach

Format

A tibble with one row per date and the following columns:

REPORT_DATE

Date of observation (class: Date)

AUS, NSW, VIC, QLD, SA, WA, TAS, NT, ACT

Daily 7-day breach rates for each jurisdiction

AUS_BREACH, NSW_BREACH, VIC_BREACH, QLD_BREACH, SA_BREACH, WA_BREACH, TAS_BREACH, NT_BREACH, ACT_BREACH

Binary indicators (1 = breach event occurred, 0 = none)

Source

Simulated data derived from Zachreson et al. (2022), COVID-19 in low-tolerance border quarantine systems: Impact of the Delta variant of SARS-CoV-2. Science Advances. Available at https://www.science.org/doi/10.1126/sciadv.abm3624.

Details

This dataset includes raw simulated data showing the 7-day breach rates of COVID-19 (or other potential viruses) within Australian quarantine systems. It provides state-level and national-level daily observations of breach probabilities under different quarantine system configurations.

The data was extracted from sheet 9 of data-raw/covid.xlsx, which is derived from simulations presented in Zachreson et al. (2022). The original study is available at https://www.science.org/doi/10.1126/sciadv.abm3624.

This dataset can be used to analyse the susceptibility of any virus given the quarantine systems implemented across different Australian jurisdictions.

Examples

data(breach)
head(breach)
#> # A tibble: 6 × 14
#>   REPORT_DATE           AUS   NSW NSW_BREACH   VIC VIC_BREACH   QLD QLD_BREACH
#>   <dttm>              <dbl> <dbl>      <dbl> <dbl>      <dbl> <dbl>      <dbl>
#> 1 2020-10-27 00:00:00  12     6.6         NA     0         NA   0.4         NA
#> 2 2020-10-28 00:00:00  11.7   6.4         NA     0         NA   0.3         NA
#> 3 2020-10-29 00:00:00  10.9   5.7         NA     0         NA   0.6         NA
#> 4 2020-10-30 00:00:00   9.9   5.6         NA     0         NA   0.1         NA
#> 5 2020-10-31 00:00:00   9.7   5.3         NA     0         NA   0.3         NA
#> 6 2020-11-01 00:00:00   8.6   4.9         NA     0         NA   0.4         NA
#> # ℹ 6 more variables: WA <dbl>, WA_BREACH <dbl>, SA <dbl>, SA_BREACH <dbl>,
#> #   NT <dbl>, NT_BREACH <lgl>